requirements.txt
azure-search-documents
openai
PyPDF2
python-dotenv